+++ app/src/main/java/com/rama/mako/activities/AboutActivity.kt
@@ -1,41 +1,11 @@
 package com.rama.mako.activities
 
-import android.content.Intent
-import android.net.Uri
-import android.os.Bundle
-import android.view.View
-import android.widget.TextView
-import com.rama.mako.CsActivity
+import com.rama.bohio.activity.BohioAboutActivity
 import com.rama.mako.R
-import com.rama.bohio.R as BohioR
 
-class AboutActivity : CsActivity() {
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-        setContentView(R.layout.view_about)
-
-        val root = findViewById<View>(R.id.about_root)
-        applyEdgeToEdgePadding(root)
-        applyCurrentTheme(root)
-
-        val closeButton = findViewById<View>(R.id.close_button)
-        closeButton.setOnClickListener {
-            finish()
-        }
-
-        val discordButton = findViewById<View>(R.id.discord_button)
-        discordButton.setOnClickListener {
-            val intent = Intent(
-                Intent.ACTION_VIEW,
-                Uri.parse(getString(BohioR.string.discord_url))
-            )
-            startActivity(intent)
-        }
-
-        val version = packageManager.getPackageInfo(packageName, 0).versionCode
-        val nameView = findViewById<TextView>(R.id.name_version)
-        nameView.text = getString(BohioR.string.name_version, getString(R.string.app_name), version)
-    }
+class AboutActivity : BohioAboutActivity() {
+    override val appIconRes = R.drawable.mako
+    override val appDescriptionRes = R.string.app_desc
+    override val appNameRes = R.string.app_name
+    override val appClaimsArrayRes = R.array.app_claims
 }